fix(web): address coderabbit review — preserve draft on plain paste, extract cleanChips to shared module
CR#1 (functional): onPaste now only commits when pasted text contains
delimiters; plain-text paste merges into draft so user can keep editing.

CR#2 (maintainability): cleanChips extracted to clean-chips.mjs (plain JS),
imported by both explore.ts and test-clean-chips.mjs — eliminates drift risk.

CR#3 (code quality): removed exploratory commentary from mixed-delimiters test.
