+++ web/src/lib/explore.ts
@@ -88,27 +88,11 @@ export type ScanEvent =
   | { kind: "error"; message: string }
   | { kind: "done"; count: number; offers: DiscoveredOffer[]; cost?: { tokens: number; usd: number } };
 
-const CHIP_CAP = 16;
-
-/** Trim, drop empties, de-dupe case-insensitively, cap length. */
-export function cleanChips(v: unknown): string[] {
-  if (v == null) return [];
-  const arr = Array.isArray(v) ? v : [v];
-  const seen = new Set<string>();
-  const out: string[] = [];
-  for (const item of arr) {
-    if (typeof item !== "string") continue;
-    const k = item.trim();
-    if (!k) continue;
-    if (!/[\p{L}\p{N}]/u.test(k)) continue; // drop punctuation-only junk (e.g. a stray "*")
-    const key = k.toLowerCase();
-    if (seen.has(key)) continue;
-    seen.add(key);
-    out.push(k);
-    if (out.length >= CHIP_CAP) break;
-  }
-  return out;
-}
+// cleanChips is defined in clean-chips.mjs (plain JS) so it can be shared
+// with the test suite without a TypeScript runner. Import for internal use
+// and re-export for external consumers (filter-builder.tsx, etc.).
+import { cleanChips } from "./clean-chips.mjs";
+export { cleanChips };
 
 function clampNum(v: unknown, lo: number, hi: number, fallback: number): number {
   const n = Number(v);
